Feature preprocessing for inference
Now, partitioning saves 2-stage masks to recreate preprocessing for reaction 1D feature arrays, as well as substrate masking.
It goes as follows:

1. amine descriptors are masked, and concatenated with masked bromide descriptors, catalyst, solvent, and base for each reaction...
this is called assemble_desc_for_inference_mols
2. that array is saved
3. a specific set of models is selected with which predictions will be made, and those models' project's partition directory
contains saved masks that can be used to recreate properly oriented feature arrays from the reaction components. The details
of preprocessing is flexible, so that different variance thresholds, for example, can be applied for different modeling experimetns
4. using the saved masks, a new function called preprocess_prophetic_features is called, which will build a tf_organizer instance
which "knows" where the prophetic feature array locations are as an attribute ("prophetic_features"), and these are organized
in the same order as partitions. this means that a driver instance can now be used, and the prophetic features can be accessed
in a pretty painless way (using a string filepath, just like the other partitions are handled) that should scale with large models/requests.